+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -1089,6 +1089,18 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     }
 
     @Test
+    @TestMetadata("NestedBindingContainersRespectScopes.kt")
+    public void testNestedBindingContainersRespectScopes() {
+      run("NestedBindingContainersRespectScopes.kt");
+    }
+
+    @Test
+    @TestMetadata("NestedBindingContainersRespectScopesAcrossModules.kt")
+    public void testNestedBindingContainersRespectScopesAcrossModules() {
+      run("NestedBindingContainersRespectScopesAcrossModules.kt");
+    }
+
+    @Test
     @TestMetadata("PrivateBindsProperty.kt")
     public void testPrivateBindsProperty() {
       run("PrivateBindsProperty.kt");